I realized, that we must defer the evaluation of nested components
There is this simple example where we create a component and store it in
a variable and then either conditionally use it or use it multiple
times.

Either way, that can't work if we want to use the moment of creation to
do this.  Instead, we need to call the 'Root()' component method with
it's state and only later call the nested components with their state.