Try to get React to confuse the state
I think I had this working, but now it seems to be broken again.  Not
sure if this is inconsitent or if my eyes just fooled me.  However, the
underlying problem exists.

What I did not expect is that React's heuristic uses the 'false' value
to guess that the component was rendered conditionally.