Resolve issue with event handlers triggering multiple times
I am not entirely sure, why this fixes the issue but it seems to be
resolved.  The underlying issue was, that I was iterating through the
wrong list to remove the event handlers.

Using the correct list didn't seem to work either though.  I just
removed the list entirely since it was redundant.